+++ app/Http/Controllers/Admin/SemesterController.php
@@ -3,89 +3,72 @@
 namespace App\Http\Controllers\Admin;
 
 use App\Models\Semester;
+use Illuminate\Support\Facades\Log;
 use App\Http\Controllers\Controller;
+use Illuminate\Support\Facades\Gate;
+use App\Http\Requests\SemesterRequest;
 use Illuminate\Database\QueryException;
 use App\Http\Resources\SemesterCollection;
-use App\Http\Requests\StoreSemesterRequest;
 use App\Http\Requests\UpdateSemesterRequest;
 
 class SemesterController extends Controller
 {
-    /**
-     * Display a listing of the resource.
-     */
-    public function index()
-    {
-
-        // get program id from the request
-
-        $programId = request()->input('program_id');
-
-        try {
-            return response()->json(new SemesterCollection(Semester::all()->where('program_id', $programId)->sortByDesc('updated_at')), 200); // 200 OK
-        } catch (QueryException $exception) {
-            return response()->json(['error' => 'Database error'.$exception->getMessage()], 500);
-        }
-
-    }
-
-    /**
-     * Show the form for creating a new resource.
-     */
-    public function create()
-    {
-        //
-    }
+    public const ONLY = ['store', 'update', 'destroy'];
 
     /**
      * Store a newly created resource in storage.
      */
-    public function store(StoreSemesterRequest $request)
+    public function store(SemesterRequest $request)
     {
-        // write store method like Department store method
+        $attributes = $request->validated();
+        $response   = Gate::inspect('create', Semester::class);
+        $message    = '';
         try {
-            $semester = Semester::create($request->all());
 
-            return response()->json($semester, 201); // 201 Created
+            if ($response->allowed()) {
+                Semester::create($attributes);
+
+                return back()->with('success', 'Semester successfully created');
+            } else {
+                $message = $response->message();
+            }
+
         } catch (QueryException $exception) {
-            return response()->json(['error' => 'Constraint violation or other database error'.$exception->getMessage()], 422);
-        }
-    }
+            $logData = ["message" => $exception->getMessage(), 'file' => $exception->getFile(), 'line' => $exception->getLine()];
+            Log::channel('semesters')->error('QueryException', $logData);
 
-    /**
-     * Display the specified resource.
-     */
-    public function show(Semester $semester)
-    {
-        // write show method like Day show method
-        if (! $semester) {
-            return response()->json(['message' => 'Semester not found'], 404);
+            $message = 'Database error 👉 Something went wrong!';
         }
 
-        return response()->json($semester);
-    }
-
-    /**
-     * Show the form for editing the specified resource.
-     */
-    public function edit(Semester $semester)
-    {
-        //
+        return back()->withErrors(['message' => $message]);
     }
 
     /**
      * Update the specified resource in storage.
      */
-    public function update(UpdateSemesterRequest $request, Semester $semester)
+    public function update(SemesterRequest $request, Semester $semester)
     {
-        // write update method like Day update method
+        $attributes = $request->validated();
+        $response   = Gate::inspect('update', $semester);
+        $message    = '';
         try {
-            $semester->update($request->all());
 
-            return response()->json($semester, 200); // 200 OK
+            if ($response->allowed()) {
+                $semester->update($attributes);
+
+                return back()->with('success', 'Semester successfully updated');
+            } else {
+                $message = $response->message();
+            }
+
         } catch (QueryException $exception) {
-            return response()->json(['error' => 'Database error'.$exception->getMessage()], 500);
+            $logData = ["message" => $exception->getMessage(), 'file' => $exception->getFile(), 'line' => $exception->getLine()];
+            Log::channel('semesters')->error('QueryException', $logData);
+
+            $message = 'Database error 👉 Something went wrong!';
         }
+
+        return back()->withErrors(['message' => $message]);
     }
 
     /**
@@ -93,12 +76,25 @@ class SemesterController extends Controller
      */
     public function destroy(Semester $semester)
     {
+        $response = Gate::inspect('delete', $semester);
+        $message  = '';
         try {
-            $semester->delete();
 
-            return response()->json(['semester' => $semester,  'message' => 'Resource successfully deleted'], 200);
+            if ($response->allowed()) {
+                $semester->delete();
+
+                return back()->with('success', 'Semester successfully deleted');
+            } else {
+                $message = $response->message();
+            }
+
         } catch (QueryException $exception) {
-            return response()->json(['error' => 'Database error'.$exception->getMessage()], 500);
+            $logData = ["message" => $exception->getMessage(), 'file' => $exception->getFile(), 'line' => $exception->getLine()];
+            Log::channel('semesters')->error('QueryException', $logData);
+
+            $message = 'Database error 👉 Something went wrong!';
         }
+
+        return back()->with('error', $message);
     }
 }
